requirements.txt
alabaster==0.7.12
alembic==1.4.2
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
bootstrap-py==0.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
defusedxml==0.6.0
docutils==0.16
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gitdb==4.0.4
GitPython==3.1.1
idna==2.9
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
packaging==20.3
pguard==0.4.0
Pillow==7.1.1
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
smmap==3.0.2
snowballstemmer==2.0.0
Sphinx==3.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.16
urllib3==1.25.8
Werkzeug==1.0.1
WTForms==2.2.1
